Added Swagger docs for users routes

+++ api/migrations/20260625155233_add_id_to_user_favorite_plants_table.js
@@ -1,19 +1,26 @@
+
 /**
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.up = async function (knex) {
-  await knex.schema.alterTable("users_favorite_plants", (table) => {
-    table.increments("id").primary(); // add new PK
-    table.unique(["user_id", "plant_id"]); // enforce uniqueness
-  });
+export const up = async function (knex) {
+  const exists = await knex.schema.hasColumn("users_favorite_plants", "id");
+
+  if (!exists) {
+    await knex.schema.alterTable("users_favorite_plants", (table) => {
+      table.increments("id").primary();
+    });
+  }
 };
 
-exports.down = async function (knex) {
-  await knex.schema.alterTable("users_favorite_plants", (table) => {
-    table.dropUnique(["user_id", "plant_id"]);
-    table.dropColumn("id");
-  });
+export const down = async function (knex) {
+  const exists = await knex.schema.hasColumn("users_favorite_plants", "id");
+
+  if (exists) {
+    await knex.schema.alterTable("users_favorite_plants", (table) => {
+      table.dropColumn("id");
+    });
+  }
 };
 
 /**
